feat: refactor champion detail page
